use correct project page and add TODOs
